MinMax Method

MinMax is used to fit a linear regression model to symbolic interval-valued variables based on the MinMax method (Lima Neto and De Carvalho, 2008).

MinMax(formula1, formula2, data, ...)

Arguments

formula1
an object of class "formula": a symbolic description of the model to be fitted.

formula2
an object of class "formula": a symbolic description of the model to be fitted.

data
an optional data frame containing the variables in the model.

…
other arguments.

Details

The Min-Max Method suggests to estimate the lower and upper bounds of the intervals using different vectors of parameters. This is equivalent to supposing independence between the values of lower and upper bounds of the intervals. The MinMax Method fits two independent linear regression models on the lower and upper bounds of the intervals, respectively, and minimizes the error of the lower bounds plus the error of the upper bounds.

Value

MinMax returns an object of class "MinMax" including at least the following elements:

coefficients.l

a named vector of coefficients for the Minimum explanatory variables.

coefficients.u

a named vector of coefficients for the Maximum explanatory variables.

sigma.l

an estimate of standard deviation for the Minimum response variable

sigma.u

an estimate of standard deviation for the Maximum response variable

df.l

the degrees of freedom for the lower residuals

df.u

the degrees of freedom for the upper residuals

fitted.values.l

the fitted values for the lower interval bound.

fitted.values.u

the fitted values for the upper interval bound.

residuals.l

the ordinary residuals for the lower interval bound.

residuals.u

the ordinary residuals for the upper interval bound.

References

Billard, L. and Diday, E. (2000) Regression analysis for interval-valued data. Data Analysis, Classification and Related Methods: Proceedings of the Seventh Conference of the International Federation of Classification Societies, Springer-Verlag, pp. 369-374.

Lima Neto, E.A. and De Carvalho, F.A.T. (2008). Centre and range method to fitting a linear regression model on symbolic interval data. Computational Statistics and Data Analysis, 52, 1500--1515.

Note

formula1 must contain the lower limit of the symbolic interval-valued variables. formula2 contain the upper limit of the symbolic interval-valued variables.

See also

Examples

data("Cardiological.MinMax", package = "iRegression") ## see Billard, L. and Diday, E. (2000)
ex.MinMax <- MinMax(PulseMin~SystMin+DiastMin,PulseMax~SystMax+DiastMax,data=Cardiological.MinMax)
ex.MinMax

#> Call:
#> MinMax.formula(formula1 = PulseMin ~ SystMin + DiastMin, formula2 = PulseMax ~ 
#>     SystMax + DiastMax, data = Cardiological.MinMax)
#> 
#> $coefficients.l
#> (Intercept)     SystMin    DiastMin 
#>  22.5765996   0.2654480   0.1952525 
#> 
#> $coefficients.u
#> (Intercept)     SystMax    DiastMax 
#>  34.0809692   0.3061762   0.1089141 
#> 
#> $sigma.l
#> [1] 12.18673
#> 
#> $sigma.u
#> [1] 11.3557
#> 
#> $df.l
#> [1] 8
#> 
#> $df.u
#> [1] 8
#> 
#> $fitted.values.l
#>  [1] 56.22955 60.13460 77.31205 67.39608 56.22955 72.70504 65.83883 71.92403
#>  [9] 65.44356 76.78115 67.00558
#> 
#> $fitted.values.u
#>  [1]  72.32258  83.68614 100.08410  89.32071  72.32258  95.04971  81.03570
#>  [8]  92.87143 104.23500 101.17324  90.89881
#> 
#> $residuals.l
#>  [1] -12.22954635  -0.13459560 -21.31204685   2.60391898  -2.22954635
#>  [6]  -2.70504183  -2.83882675   0.07596802  10.55644360   9.21884923
#> [11]  18.99442390
#> 
#> $residuals.u
#>  [1]  -4.3225766 -11.6861447 -10.0840963  22.6792873  -0.3225766   4.9502872
#>  [7]  -6.0357032   7.1285689  -6.2349992  -5.1732371   9.1011902
#>