Constrained Centre and Range Method

ccrm is used to fit a linear regression model to symbolic interval-valued variables based on the inequality constraints over the range variables (Lima Neto and De Carvalho, 2010).

ccrm(formula1, formula2, data, ...)

Arguments

formula1
an object of class "formula": the description of the first model to be fitted.

formula2
an object of class "formula": the description of the second model to be fitted.

data
an optional data frame containing the variables in the model.

…
other arguments.

Details

The Constrained Centre and Range method (CCRM) was proposed by Lima Neto and De Carvalho (2010) and fits two independent linear regression models on the midpoint and range of the intervals. In the Constrained Centre and Range Method, the estimative of the parameters of the range's model is based on inequality constraints. There is no constraints over the parameters estimates for the midpoint regression equation. The aim is to guarantee mathematical coherence between the predicted values of the lower and upper bounds of the response interval-valued variable Y, i.e., yL < yU.

Value

ccrm returns an object of class "ccrm" including at least the following elements:

coefficients.C

a named vector of coefficients for the Centre's explanatory variables.

coefficients.R

a named vector of coefficients for the Range's explanatory variables.

sigma.C

an estimative of the standard deviation for the Centre's response variable.

sigma.R

an estimative of the standard deviation for the Range's response variable.

df.C

the degrees of freedom for the Centre residuals

df.R

the degrees of freedom for the Range residuals

fitted.values.l

the fitted values for the lower interval bound.

fitted.values.u

the fitted values for the upper interval bound.

residuals.l

the ordinary residuals for the lower interval bound.

residuals.u

the ordinary residuals for the upper interval bound.

References

Lima Neto, E.A. and De Carvalho, F.A.T. (2010). Constrained linear regression models for symbolic interval-valued variables. Computational Statistics and Data Analysis, 54, 333--347.

Note

formula1 must contain the midpoint of the symbolic interval-valued variables. formula2 contain the range (upper limit minus lower limit) of the symbolic interval-valued variables.

See also

Examples

data("Cardiological.CR", package = "iRegression")
ex.ccrm <- ccrm(PulseC~SystC+DiastC,PulseR~SystR+DiastR,data=Cardiological.CR)
ex.ccrm

#> Call:
#> ccrm.formula(formula1 = PulseC ~ SystC + DiastC, formula2 = PulseR ~ 
#>     SystR + DiastR, data = Cardiological.CR)
#> 
#> $coefficients.C
#> (Intercept)       SystC      DiastC 
#>  21.1708061   0.3288879   0.1698512 
#> 
#> $coefficients.R
#> (Intercept)        <NA>        <NA> 
#>  17.9555967   0.0000000   0.2072155 
#> 
#> $sigma.C
#> [1] 9.516986
#> 
#> $sigma.R
#> [1] 11.38413
#> 
#> $df.C
#> [1] 8
#> 
#> $df.R
#> [1] 8
#> 
#> $fitted.values.l
#>  [1] 51.55627 59.88662 79.91486 66.69788 51.55627 72.90938 62.09638 72.52889
#>  [9] 72.66849 79.39915 67.78582
#> 
#> $fitted.values.u
#>  [1]  73.65618  81.98652  99.94261  90.45551  73.65618  97.08144  82.12413
#>  [8]  93.38551  98.91271 101.49905  90.30016
#> 
#> $residuals.l
#>  [1]  -7.5562732   0.1133848 -23.9148555   3.3021242   2.4437268  -2.9093816
#>  [7]   0.9036178  -0.5288918   3.3315125   6.6008542  18.2141818
#> 
#> $residuals.u
#>  [1] -5.6561807 -9.9865227 -9.9426076 21.5444923 -1.6561807  2.9185555
#>  [7] -7.1241343  6.6144939 -0.9127058 -5.4990532  9.6998432
#>