Centre and Range Method

crm is used to fit a linear regression model to symbolic interval-valued variables based on the Centre and Range method (Lima Neto and De Carvalho, 2008).

crm(formula1, formula2, data, ...)

Arguments

formula1
an object of class "formula": a symbolic description of the model to be fitted.

formula2
an object of class "formula": a symbolic description of the model to be fitted.

data
an optional data frame containing the variables in the model.

…
other arguments.

Details

In the Center Method, the estimate of the parameters beta is based only on the midpoint of the intervals. However, the Centre and Range Method proposed by Lima Neto and De Carvalho (2008) consider suitable to include both the information given by the center and by the range of an interval-valued variable on a linear regression model to improve the model prediction performance. The Centre and Range Method fits two independent linear regression models on the midpoint and range of the intervals, respectively, and minimizes the error of the midpoint plus the error of the range.

Value

cm returns an object of class "crm" including at least the following elements:

coefficients.C

a named vector of coefficients for the Centre variables.

coefficients.R

a named vector of coefficients for the Range variables.

sigma.C

an estimate of standard deviation for the Centre response variable.

sigma.R

an estimate of standard deviation for the Range response variable.

df.C

the degrees of freedom for the centre residuals

df.R

the degrees of freedom for the range residuals

fitted.values.l

the fitted mean values for the lower interval bound.

fitted.values.u

the fitted mean values for the upper interval bound.

residuals.l

the residuals for the lower interval bound (that is response minus fitted values).

residuals.u

the residuals for the upper interval bound (that is response minus fitted values).

References

Billard, L. and Diday, E. (2000) Regression analysis for interval-valued data. Data Analysis, Classification and Related Methods: Proceedings of the Seventh Conference of the International Federation of Classification Societies, Springer-Verlag, pp. 369-374.

Lima Neto, E.A. and De Carvalho, F.A.T. (2008). Centre and range method to fitting a linear regression model on symbolic interval data. Computational Statistics and Data Analysis, 52, 1500--1515.

Note

formula1 must contain the midpoint of the symbolic interval-valued variables. formula2 contain the range (upper limit minus lower limit) of the symbolic interval-valued variables.

See also

Examples

data("Cardiological.CR", package = "iRegression")
ex.crm <- crm(PulseC~SystC+DiastC,PulseR~SystR+DiastR,data=Cardiological.CR)
ex.crm

#> Call:
#> crm.formula(formula1 = PulseC ~ SystC + DiastC, formula2 = PulseR ~ 
#>     SystR + DiastR, data = Cardiological.CR)
#> 
#> $coefficients.C
#> (Intercept)       SystC      DiastC 
#>  21.1708061   0.3288879   0.1698512 
#> 
#> $coefficients.R
#> (Intercept)       SystR      DiastR 
#>  20.2149263  -0.1466956   0.3480081 
#> 
#> $sigma.C
#> [1] 9.516986
#> 
#> $sigma.R
#> [1] 11.05429
#> 
#> $df.C
#> [1] 8
#> 
#> $df.R
#> [1] 8
#> 
#> $fitted.values.l
#>  [1] 49.75216 60.28294 81.01514 65.94425 49.75216 71.86826 63.19667 72.61411
#>  [9] 74.59080 79.94216 68.04135
#> 
#> $fitted.values.u
#>  [1]  75.46029  81.59020  98.84232  91.20914  75.46029  98.12256  81.02385
#>  [8]  93.30028  96.99040 100.95604  90.04463
#> 
#> $residuals.l
#>  [1]  -5.7521614  -0.2829376 -25.0151405   4.0557542   4.2478386  -1.8682634
#>  [7]  -0.1966672  -0.6141137   1.4092028   6.0578362  17.9586519
#> 
#> $residuals.u
#>  [1] -7.460293 -9.590200 -8.842323 20.790862 -3.460293  1.877437 -6.023849
#>  [8]  6.699716  1.009604 -4.956035  9.955373
#>