Regression methods for interval-valued variables

This package contains some important regression methods for interval-valued variables. For each method, it is available the fitted values, residuals and some goodness-of-fit measures.

Details

Package:

iRegression

Type:	Package

Version:

1.1

Date:	2011-06-01

License:

GPL (>= 2)

LazyLoad:	yes

Some available functions: cm, MinMax, crm, ccrm, bivar

References

Billard, L. and Diday, E. (2000) Regression analysis for interval-valued data. Data Analysis, Classification and Related Methods: Proceedings of the Seventh Conference of the International Federation of Classification Societies, Springer-Verlag, pp. 369-374.

Lima Neto, E.A. and De Carvalho, F.A.T. (2008). Centre and range method to fitting a linear regression model on symbolic interval data. Computational Statistics and Data Analysis, 52, 1500--1515.

Lima Neto, E.A. and De Carvalho, F.A.T. (2010). Constrained linear regression models for symbolic interval-valued variables. Computational Statistics and Data Analysis, 54, 333--347.

Lima Neto, E. A., Cordeiro, G. and De Carvalho, F.A.T. (2011). Bivariate symbolic regression models for interval-valued variables. Journal of Statistical Computation and Simulation (Print), 81, 1727--1744.