Summarizing MinMax Method Fits

summary method for class MinMax.

# S3 method for MinMax
summary(object, ...)

Arguments

object
an object of class "MinMax", usually, a result of a call to MinMax

…
other arguments.

Value

The function summary.MinMax returns the following elements, given an object of the class "MinMax",

Coef.L

a named vector of coefficients for the Min explanatory variables.

Coef.U

a named vector of coefficients for the Max explanatory variables.

RMSE.l

root mean square error for the lower bound.

RMSE.u

root mean square error for the upper bound.

References

Billard, L. and Diday, E. (2000) Regression analysis for interval-valued data. Data Analysis, Classification and Related Methods: Proceedings of the Seventh Conference of the International Federation of Classification Societies, Springer-Verlag, pp. 369-374.

Lima Neto, E.A. and De Carvalho, F.A.T. (2008). Centre and range method to fitting a linear regression model on symbolic interval data. Computational Statistics and Data Analysis, 52, 1500--1515.

See also

Examples

##-- Continuing the  MinMax() example:
data("Cardiological.MinMax", package = "iRegression")
ex.MinMax <- MinMax(PulseMin~SystMin+DiastMin,PulseMax~SystMax+DiastMax,data=Cardiological.MinMax)
ex.sum <- summary(ex.MinMax)
ex.sum

#> Call:
#> MinMax.formula(formula1 = PulseMin ~ SystMin + DiastMin, formula2 = PulseMax ~ 
#>     SystMax + DiastMax, data = Cardiological.MinMax)
#> 
#>             Estimate.L   StdErr.L
#> (Intercept) 22.5765996 23.6212767
#> SystMin      0.2654480  0.1578517
#> DiastMin     0.1952525  0.1625865
#> 
#> RMSE.L:
#> [1] 10.39287
#> 
#>             Estimate.U   StdErr.U
#> (Intercept) 34.0809692 21.0028892
#> SystMax      0.3061762  0.1089318
#> DiastMax     0.1089141  0.1658257
#> 
#> RMSE.U:
#> [1] 9.68417